Update README.md
This project is a detailed CV of Itihas Verma, a Junior Frontend Developer. The CV is created using HTML and CSS, and it includes personal information, skills, education, and work experience. The page is designed to be responsive and includes social media links with icons from Font Awesome and FlatIcon.